refactor(view-mode): mueve modules/TypeOfView a shared/preferences/view-mode
