+++ src/shared/preferences/view-mode/index.ts
@@ -1,0 +1,9 @@
+export {
+	TypeOfView,
+	defaultView,
+	isValidTypeOfView,
+	typeOfViewLocalStorageKey,
+} from './model/typeOfView'
+export { useTypeOfView } from './hooks/useTypeOfView'
+export { ToggleTypeOfView } from './ui/ToggleTypeOfView'
+export { CheckView } from './ui/CheckView'
