+++ src/shared/preferences/view-mode/model/typeOfView.ts
