+++ src/views/BoardPage.tsx
@@ -11,7 +11,7 @@ import { useBoardQuery } from '@/modules/board/hooks/useBoardQuery'
 import { USER_IS_IN } from '@/shared/ui/organisms/userIsIn'
 import { ListView } from '@/modules/TaskBoard/components/ListView'
 import { TableView } from '@/modules/TaskBoard/components/TableView'
-import { useTypeOfView } from '@/modules/TypeOfView/useTypeOfView'
+import { useTypeOfView } from '@/shared/preferences/view-mode'
 import { NoteInput } from '@/features/notes'
 import { useTaskBoardQuery } from '@/modules/TaskBoard/hooks/useTaskBoardQuery'
 import { useReminder } from '@/modules/TaskBoard/components/Reminder/hooks/useReminder'
